Fix：リファクタリング（コンポーネント、ロジック）
