+++ src/components/elements/google/index.js
@@ -1,0 +1,1 @@
+export { GoogleMap } from './GoogleMap'
