+++ src/router/Router.jsx
@@ -1,0 +1,12 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import { Top } from '@/components/pages'
+
+export const Router = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Top />} />
+      </Routes>
+    </BrowserRouter>
+  )
+}
